+++ source/Application/Controller/Admin/AdminNewsletter.php
@@ -7,9 +7,14 @@
 
 namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
 
+use OxidEsales\EshopCommunity\Application\Model\NewsletterRecipients;
+use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
+use OxidEsales\EshopCommunity\Internal\Framework\FileSystem\FileGenerator\Bridge\FileGeneratorBridgeInterface;
+use OxidEsales\EshopCommunity\Internal\Framework\FileSystem\HeaderGenerator\Bridge\HeaderGeneratorBridgeInterface;
+
 /**
  * Admin newsletter manager.
- * Returns template, that arranges template ("newsletter_main.tpl") to frame.
+ * Returns template, that arranges template ("newsletter.tpl") to frame.
  * Admin Menu: Customer Info -> Newsletter.
  */
 class AdminNewsletter extends \OxidEsales\Eshop\Application\Controller\Admin\AdminController
@@ -20,4 +25,31 @@ class AdminNewsletter extends \OxidEsales\Eshop\Application\Controller\Admin\Adm
      * @var string
      */
     protected $_sThisTemplate = 'newsletter.tpl';
+
+    public function export(): void
+    {
+        $this->setCSVHeader();
+        $this->generateCSV((new NewsletterRecipients())->getNewsletterRecipients());
+        exit();
+    }
+
+    private function setCSVHeader(): void
+    {
+        $container = ContainerFactory::getInstance()->getContainer();
+        $filename = "Export_recipients_" . date("Y-m-d") . ".csv";
+        $csvHeader = $container->get(HeaderGeneratorBridgeInterface::class);
+        $csvHeader->generate($filename);
+    }
+
+    /**
+     * @param array $data
+     *
+     * @return void
+     */
+    public function generateCSV(array $data): void
+    {
+        $container = ContainerFactory::getInstance()->getContainer();
+        $csvGenerator = $container->get(FileGeneratorBridgeInterface::class);
+        $csvGenerator->generate("php://output", $data);
+    }
 }