+++ source/Internal/Framework/FileSystem/FileGenerator/FileGeneratorInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\EshopCommunity\Internal\Framework\FileSystem\FileGenerator;
+
+interface FileGeneratorInterface
+{
+    public function generate(string $filename, array $data): void;
+}