+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Example dist files for translations in Application/translations folder [PR-826](https://github.com/OXID-eSales/oxideshop_ce/pull/826)
 - Support for single language map file in translations directory [PR-449](https://github.com/OXID-eSales/oxideshop_ce/pull/449).
 - Setting tracking URL per Shipping Method
+- Add functionality to export newsletter recipients in Admin area
 
 ### Changed
 - Source/Application/views/admin/tpl/shop_license.tpl